Refactor SesAuthHandler and tidy AWSSDKUtils for code clarity
Modified the SesAuthHandler class, incorporating clearer variable naming and consistent use of var keyword for variable declaration in the codebase, contributing to cleaner, more maintainable code. The usage of authorization scheme and SHA256 hash generator was improved for clearer communication of intent and added AwsContentSha256HeaderName for more consistent naming.

Converted types to nullable reference types, and improved method readability with the more modern 'is not null' and is null' checks.

Also performed housekeeping on the AWSSDKUtils class, by correcting the naming convention and appropriately marking properties with correct access modifiers. Specifically, correct spelling on "rfcNumber" param description and modifications to UrlEncode method for better performance.