Update email provider tests and refactor AmazonSesParams
This commit overhauls the email provider test suites and adds a ConfigurationFixture for shared configurations. A refactoring of AmazonSesParams has also been carried out to add validation to the constructors, ensuring required properties are not empty. The amendments also entail renaming of "MailEase.Test" to "MailEase.Tests" and activation of email validation in AzureCommunicationEmailProvider.