Add test output helper to SharedKeyAuthHandlerTests
A private instance of ITestOutputHelper has been added to the SharedKeyAuthHandlerTests class. It is primarily used to write additional information when tests are executed, helping in the process of testing, debugging, and diagnosing in case of failure or issues. The constructor for SharedKeyAuthHandlerTests has been updated to take in an instance of ITestOutputHelper.