Add Apache and MIT license files
Two license files have been added to the project. The first, `STOWAGE_LICENSE`, is the Apache License Version 2.0, which sets terms and conditions for the use, reproduction, and distribution of the work. The second, `AZURE_SDK_FOR_NET_LICENSE.txt`, is the MIT license, permitting free use, modification, and distribution of the Microsoft software under certain conditions.